// src/app/pages/characters/models/icharacterDetail.ts
export interface ICharacterDetail {
    name: string;
    image: string;
    species: string;
    gender: string;
    status: string;

}

export interface IOriginDetail {
    name: string
}

export interface ILocationDetail {
    name: string;   
}

export interface INumberOfCharacters {
    number: number;
}